API구현/에러처리 완료

+++ routes/posts.js
@@ -5,43 +5,53 @@ const Posts = require('../schemas/post');
 
 //게시글 조회 API
 router.get('/', async (req, res) => {
-    const posts = await Posts.find();
+    try {
+        const posts = await Posts.find();
 
-    //가공해서 넣을 새로운 posts(postId, user, title, createdAt)
-    const newPosts = [];
-    posts.forEach((v) => {
-        newPosts.push({
-            postId: String(v['_id']),
-            user: v['user'],
-            title: v['title'],
-            createdAt: v['createdAt'],
-        });
-    })
+        //가공해서 넣을 새로운 posts(postId, user, title, createdAt)
+        const newPosts = [];
+        posts.forEach((v) => {
+            newPosts.push({
+                postId: String(v['_id']),
+                user: v['user'],
+                title: v['title'],
+                createdAt: v['createdAt'],
+            });
+        })
+        newPosts.sort((a, b) => (a.createdAt > b.createdAt ? -1 : 1));
 
-    res.status(200).json({ data: newPosts });
+        res.status(200).json({ data: newPosts });
+    } catch (err) {
+        res.status(400).send({ message: '잘못된 요청입니다.' });
+    }
 });
-
 
 //게시글 작성 API
 router.post('/', async (req, res) => {
-    //바디에 들어온 변수들 객체분해할당
-    const { user, password, title, content } = req.body;
+    try {
+        //바디에 들어온 변수들 객체분해할당
+        const { user, password, title, content } = req.body;
 
-    //현재시각 및 한국시간으로 변환
-    let now = new Date();
-    let createdAt = new Date(now.getTime() - (now.getTimezoneOffset() * 60000)).toISOString();
 
-    //db저장
-    const createdPost = await Posts.create({ user, password, title, content, createdAt });
+        //현재시각 및 한국시간으로 변환
+        let now = new Date();
+        let createdAt = new Date(now.getTime() - (now.getTimezoneOffset() * 60000)).toISOString();
 
-    res.status(200).json({ message: '게시글을 생성하였습니다.', post: createdPost });
+        //db저장
+        //이 부분에서 req.body가 비어있으면 validationError를 발생시킴.
+        await Posts.create({ user, password, title, content, createdAt });
+
+        res.status(200).json({ message: '게시글을 생성하였습니다.' });
+    } catch (error) {
+        res.status(400).json({ message: '데이터 형식이 올바르지 않습니다.' })
+    }
 })
 
 //게시글 상세 조회 API
 router.get('/:_postId', async (req, res) => {
-    //request.params로 받은 값을 객체분해할당.
-    const { _postId } = req.params;
     try {
+        //request.params로 받은 값을 객체분해할당.
+        const { _postId } = req.params;
         //쿼리연산자를 사용하여 조회. ($eq 는 주어진 갑과 일치하는 값.)
         const post = await Posts.findOne({ _id: { $eq: _postId } });
 
@@ -55,6 +65,7 @@ router.get('/:_postId', async (req, res) => {
         }
         res.status(200).json({ data });
     } catch (message) {
+        console.log(message);
         res.status(404).json({ message: '데이터 형식이 올바르지 않습니다.' });
     }
 
@@ -62,37 +73,71 @@ router.get('/:_postId', async (req, res) => {
 
 //게시글 수정 API
 router.put('/:_postId', async (req, res) => {
-    //response, request 의 바디와 파람스로 값 가져옴.
-    const { password, title, content } = req.body;
-    const { _postId } = req.params;
+    try {
+        if (!Object.values(req.body).length || Object.values(req.body).includes("")) {
+            throw (new Error('RequestError'));
+        }
+        //response, request 의 바디와 파람스로 값 가져옴.
+        const { password, title, content } = req.body;
+        const { _postId } = req.params;
 
-    //파람스로 받은 id와 같은 post 받아옴.
-    const post = await Posts.findOne({ _id: { $eq: _postId } });
-
-    //비밀번호 맞으면 시간 새로 지정하고, 그 post를 바로 updateOne()
-    if (post['password'] === password) {
-        let now = new Date();
-        let createdAt = new Date(
-            now.getTime() - (now.getTimezoneOffset() * 60000))
-            .toISOString();
-        await post.updateOne(
-            { $set: { title, content, createdAt } }
-        );
+        //파람스로 받은 id와 같은 post 받아옴.
+        const post = await Posts.findOne({ _id: { $eq: _postId } });
+        //비밀번호 맞으면 시간 새로 지정하고, 그 post를 바로 updateOne()
+        if (post['password'] === password) {
+            let now = new Date();
+            let createdAt = new Date(
+                now.getTime() - (now.getTimezoneOffset() * 60000))
+                .toISOString();
+            await post.updateOne(
+                { $set: { title, content, createdAt } }
+            );
+            return res.status(200).json({ message: '게시글을 수정하였습니다.' });
+        } else {
+            throw new Error('PasswordError');
+        }
+    } catch (error) {
+        if (error.message == 'RequestError') {
+            res.status(400).json({ message: '데이터 형식이 올바르지 않습니다.' })
+        } else if (error.message == 'PasswordError') {
+            res.status(400).json({ message: '비밀번호가 일치하지 않습니다.' })
+        } else {
+            res.status(404).json({ message: '게시글 조회에 실패하였습니다.' })
+        }
     }
-    res.status(200).json({ message: '게시글을 수정하였습니다.' })
 })
+
+
+
 
 //게시글 삭제 API
 router.delete('/:_postId', async (req, res) => {
-    const { _postId } = req.params;
-    const { password } = req.body;
+    try {
+        if (!Object.values(req.body).length || Object.values(req.body).includes("")) {
+            throw (new Error('RequestError'));
+        }
+        //response, request 의 바디와 파람스로 값 가져옴.
+        const { _postId } = req.params;
+        const { password } = req.body;
 
-    //받아온 값들로 해당 post찾고 비밀번호 확인 후 삭제
-    const post = await Posts.findOne({ _id: { $eq: _postId } });
-    if (post['password'] === password) {
-        await post.deleteOne();
+        //파람스로 받은 id와 같은 post 받아옴.
+        const post = await Posts.findOne({ _id: { $eq: _postId } });
+        //비밀번호 맞으면 시간 새로 지정하고, 그 post를 바로 updateOne()
+        if (post['password'] === password) {
+            await post.deleteOne();
+            return res.status(200).json({ message: '게시글을 삭제하였습니다.' });
+        } else {
+            throw new Error('PasswordError');
+        }
+    } catch (error) {
+        if (error.message == 'RequestError') {
+            res.status(400).json({ message: '데이터 형식이 올바르지 않습니다.' })
+        } else if (error.message == 'PasswordError') {
+            res.status(400).json({ message: '비밀번호가 일치하지 않습니다.' })
+        } else {
+            res.status(404).json({ message: '게시글 조회에 실패하였습니다.' })
+        }
     }
-    res.json({ message: '게시글을 삭제하였습니다.' })
 })
 
 module.exports = router;
